Finalized implementation of CalcPathLengthAndCost
