+++ Source/SVONavigation/Private/PathFinding/SVOPathFindingAlgorithm.cpp
@@ -6,52 +6,6 @@
 #include "SVONavigationData.h"
 #include "SVOVolumeNavigationData.h"
 
-FSVONodeAddressWithLocation::FSVONodeAddressWithLocation( const FSVONodeAddress & node_address, const FVector & location ) :
-    NodeAddress( node_address ),
-    Location( location )
-{}
-
-FSVONodeAddressWithLocation::FSVONodeAddressWithLocation( const FSVONodeAddress & node_address, const FSVOVolumeNavigationData & bounds_navigation_data ) :
-    NodeAddress( node_address ),
-    Location( bounds_navigation_data.GetNodePositionFromAddress( node_address ) )
-{}
-
-FSVOPathFinderDebugNodeCost::FSVOPathFinderDebugNodeCost( const FSVONodeAddressWithLocation & from, const FSVONodeAddressWithLocation & to, const float cost, const bool is_closed ) :
-    From( from ),
-    To( to ),
-    Cost( cost ),
-    bIsClosed( is_closed )
-{
-}
-
-void FSVOPathFinderDebugNodeCost::Reset()
-{
-    *this = FSVOPathFinderDebugNodeCost();
-}
-
-void FSVOPathFinderDebugInfos::Reset()
-{
-    LastProcessedSingleNode.Reset();
-    ProcessedNeighbors.Reset();
-    Iterations = 0;
-    VisitedNodes = 0;
-    CurrentBestPath.ResetForRepath();
-}
-
-FSVOPathFindingParameters::FSVOPathFindingParameters( const FSVOVolumeNavigationData & volume_navigation_data, const FVector & start_location, const FVector & end_location, const FNavigationQueryFilter & nav_query_filter ) :
-    StartLocation( start_location ),
-    EndLocation( end_location ),
-    NavigationQueryFilter( nav_query_filter ),
-    QueryFilterImplementation( static_cast< const FSVONavigationQueryFilterImpl * >( NavigationQueryFilter.GetImplementation() ) ),
-    QueryFilterSettings( QueryFilterImplementation->QueryFilterSettings ),
-    HeuristicCalculator( QueryFilterSettings.HeuristicCalculator ),
-    CostCalculator( QueryFilterSettings.TraversalCostCalculator ),
-    VolumeNavigationData( volume_navigation_data )
-{
-    VolumeNavigationData.GetNodeAddressFromPosition( StartNodeAddress, StartLocation );
-    VolumeNavigationData.GetNodeAddressFromPosition( EndNodeAddress, EndLocation );
-}
-
 FSVOGraphAStar::FSVOGraphAStar( const FSVOVolumeNavigationData & graph ) :
     FGraphAStar< FSVOVolumeNavigationData, FGraphAStarDefaultPolicy, FGraphAStarDefaultNode< FSVOVolumeNavigationData > >( graph )
 {
@@ -97,7 +51,7 @@ ESVOPathFindingAlgorithmStepperStatus FSVOPathFindingAlgorithmStepper::Step( EGr
     }
 }
 
-bool FSVOPathFindingAlgorithmStepper::FillNodeAddresses( TArray< FSVONodeAddress > & node_addresses ) const
+bool FSVOPathFindingAlgorithmStepper::FillNodeAddresses( TArray< FSVOPathFinderNodeAddressWithCost > & node_addresses ) const
 {
     checkNoEntry();
     return false;
