+++ Source/SVONavigation/Public/PathFinding/SVOPathFindingAlgorithm.h
@@ -1,98 +1,10 @@
 #pragma once
 
-#include "SVONavigationPath.h"
-#include "SVONavigationTypes.h"
 #include "SVOPathFindingAlgorithmObservers.h"
+#include "SVOPathFindingAlgorithmTypes.h"
 #include "SVOVolumeNavigationData.h"
 
-#include <GraphAStar.h>
-#include <NavigationPath.h>
-
 #include "SVOPathFindingAlgorithm.generated.h"
-
-struct FSVONavigationPath;
-struct FSVONavigationQueryFilterSettings;
-class FSVONavigationQueryFilterImpl;
-struct FPathFindingQuery;
-class ASVONavigationData;
-
-struct FSVONodeAddressWithLocation
-{
-    FSVONodeAddressWithLocation() = default;
-    FSVONodeAddressWithLocation( const FSVONodeAddress & node_address, const FVector & location );
-    FSVONodeAddressWithLocation( const FSVONodeAddress & node_address, const FSVOVolumeNavigationData & bounds_navigation_data );
-
-    bool operator==( const FSVONodeAddressWithLocation & other ) const
-    {
-        return NodeAddress == other.NodeAddress;
-    }
-
-    bool operator!=( const FSVONodeAddressWithLocation & other ) const
-    {
-        return !operator==( other );
-    }
-
-    FSVONodeAddress NodeAddress;
-    FVector Location;
-};
-
-FORCEINLINE uint32 GetTypeHash( const FSVONodeAddressWithLocation & node_address_with_location )
-{
-    return GetTypeHash( node_address_with_location.NodeAddress );
-}
-
-struct FSVOPathFinderDebugNodeCost
-{
-    FSVOPathFinderDebugNodeCost() = default;
-    FSVOPathFinderDebugNodeCost( const FSVONodeAddressWithLocation & from, const FSVONodeAddressWithLocation & to, const float cost, const bool is_closed );
-
-    void Reset();
-
-    FSVONodeAddressWithLocation From;
-    FSVONodeAddressWithLocation To;
-    float Cost;
-    uint8 bIsClosed : 1;
-};
-
-USTRUCT()
-struct SVONAVIGATION_API FSVOPathFinderDebugInfos
-{
-    GENERATED_USTRUCT_BODY()
-
-    void Reset();
-
-    FSVOPathFinderDebugNodeCost LastProcessedSingleNode;
-    TArray< FSVOPathFinderDebugNodeCost > ProcessedNeighbors;
-    FSVONavigationPath CurrentBestPath;
-
-    UPROPERTY( VisibleAnywhere )
-    int Iterations;
-
-    UPROPERTY( VisibleAnywhere )
-    int VisitedNodes;
-
-    UPROPERTY( VisibleAnywhere )
-    int PathSegmentCount;
-
-    UPROPERTY( VisibleAnywhere )
-    float PathLength;
-};
-
-struct FSVOPathFindingParameters
-{
-    FSVOPathFindingParameters( const FSVOVolumeNavigationData & volume_navigation_data, const FVector & start_location, const FVector & end_location, const FNavigationQueryFilter & nav_query_filter );
-
-    FVector StartLocation;
-    FVector EndLocation;
-    const FNavigationQueryFilter & NavigationQueryFilter;
-    const FSVONavigationQueryFilterImpl * QueryFilterImplementation;
-    const FSVONavigationQueryFilterSettings & QueryFilterSettings;
-    const USVOPathHeuristicCalculator * HeuristicCalculator;
-    const USVOPathTraversalCostCalculator * CostCalculator;
-    const FSVOVolumeNavigationData & VolumeNavigationData;
-    FSVONodeAddress StartNodeAddress;
-    FSVONodeAddress EndNodeAddress;
-};
 
 enum class ESVOPathFindingAlgorithmState : uint8
 {
@@ -129,7 +41,7 @@ public:
     const FSVOGraphAStar & GetGraph() const;
 
     ESVOPathFindingAlgorithmStepperStatus Step( EGraphAStarResult & result );
-    virtual bool FillNodeAddresses( TArray< FSVONodeAddress > & node_addresses ) const;
+    virtual bool FillNodeAddresses( TArray< FSVOPathFinderNodeAddressWithCost > & node_addresses ) const;
 
 protected:
     virtual ESVOPathFindingAlgorithmStepperStatus Init( EGraphAStarResult & result ) = 0;
